# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Betting Site API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetcherStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/odds/{sport}": {
      "get": {
        "summary": "Get Odds",
        "operationId": "get_odds_api_odds__sport__get",
        "parameters": [
          {
            "name": "sport",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OddsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/props/{sport}": {
      "get": {
        "summary": "Get Props",
        "operationId": "get_props_api_props__sport__get",
        "parameters": [
          {
            "name": "sport",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OddsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/picks/{sport}": {
      "get": {
        "summary": "Get Picks",
        "operationId": "get_picks_api_picks__sport__get",
        "parameters": [
          {
            "name": "sport",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PicksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fetcher/start": {
      "post": {
        "summary": "Start",
        "operationId": "start_api_fetcher_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetcherControlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fetcher/stop": {
      "post": {
        "summary": "Stop",
        "operationId": "stop_api_fetcher_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetcherControlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/refresh/{event_id}": {
      "post": {
        "summary": "Refresh",
        "operationId": "refresh_api_refresh__event_id__post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetcherControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sports": {
      "get": {
        "summary": "Get Sports",
        "operationId": "get_sports_api_sports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookPrice": {
        "properties": {
          "bookmaker_key": {
            "type": "string",
            "title": "Bookmaker Key"
          },
          "price_american": {
            "type": "integer",
            "title": "Price American"
          },
          "point": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Point"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "bookmaker_key",
          "price_american",
          "fetched_at"
        ],
        "title": "BookPrice"
      },
      "FetcherControlResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "tiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tiers"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "polled": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Polled"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FetcherControlResponse"
      },
      "FetcherStatus": {
        "properties": {
          "last_fetch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fetch At"
          },
          "requests_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requests Used"
          },
          "requests_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requests Remaining"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "fetcher_running": {
            "type": "boolean",
            "title": "Fetcher Running",
            "default": false
          },
          "enabled_tiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled Tiers"
          }
        },
        "type": "object",
        "title": "FetcherStatus"
      },
      "Game": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "commence_time": {
            "type": "string",
            "format": "date-time",
            "title": "Commence Time"
          },
          "is_live": {
            "type": "boolean",
            "title": "Is Live",
            "default": false
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/Market"
            },
            "type": "array",
            "title": "Markets"
          },
          "stale_seconds": {
            "type": "integer",
            "title": "Stale Seconds",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "home_team",
          "away_team",
          "commence_time",
          "markets"
        ],
        "title": "Game"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Market": {
        "properties": {
          "market_key": {
            "type": "string",
            "title": "Market Key"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/MarketOutcome"
            },
            "type": "array",
            "title": "Outcomes"
          }
        },
        "type": "object",
        "required": [
          "market_key",
          "outcomes"
        ],
        "title": "Market"
      },
      "MarketGroupModel": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "main_key": {
            "type": "string",
            "title": "Main Key"
          },
          "alt_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Key"
          },
          "display": {
            "type": "string",
            "enum": [
              "moneyline",
              "spread",
              "total"
            ],
            "title": "Display",
            "default": "moneyline"
          }
        },
        "type": "object",
        "required": [
          "label",
          "main_key"
        ],
        "title": "MarketGroupModel"
      },
      "MarketOutcome": {
        "properties": {
          "outcome_name": {
            "type": "string",
            "title": "Outcome Name"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/BookPrice"
            },
            "type": "array",
            "title": "Prices"
          },
          "best_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookPrice"
              },
              {
                "type": "null"
              }
            ]
          },
          "consensus_price_american": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Price American"
          }
        },
        "type": "object",
        "required": [
          "outcome_name",
          "prices"
        ],
        "title": "MarketOutcome"
      },
      "OddsResponse": {
        "properties": {
          "games": {
            "items": {
              "$ref": "#/components/schemas/Game"
            },
            "type": "array",
            "title": "Games"
          },
          "stale_seconds": {
            "type": "integer",
            "title": "Stale Seconds"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "games",
          "stale_seconds",
          "fetched_at"
        ],
        "title": "OddsResponse"
      },
      "Pick": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tier": {
            "$ref": "#/components/schemas/PickTier"
          },
          "game_label": {
            "type": "string",
            "title": "Game Label"
          },
          "market_label": {
            "type": "string",
            "title": "Market Label"
          },
          "pick_side": {
            "type": "string",
            "title": "Pick Side"
          },
          "odds_american": {
            "type": "integer",
            "title": "Odds American"
          },
          "best_book": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Book"
          },
          "stake_units": {
            "type": "number",
            "title": "Stake Units"
          },
          "probability_pct": {
            "type": "number",
            "title": "Probability Pct"
          },
          "market_probability_pct": {
            "type": "number",
            "title": "Market Probability Pct"
          },
          "edge_pct": {
            "type": "number",
            "title": "Edge Pct"
          },
          "stats": {
            "items": {
              "$ref": "#/components/schemas/PickStat"
            },
            "type": "array",
            "title": "Stats"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "agent_key": {
            "type": "string",
            "title": "Agent Key",
            "default": "baseball-agents"
          },
          "agent_record_30d": {
            "type": "string",
            "title": "Agent Record 30D",
            "default": ""
          },
          "commence_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commence Time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tier",
          "game_label",
          "market_label",
          "pick_side",
          "odds_american",
          "stake_units",
          "probability_pct",
          "market_probability_pct",
          "edge_pct",
          "reasoning"
        ],
        "title": "Pick"
      },
      "PickStat": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "PickStat"
      },
      "PickTier": {
        "type": "string",
        "enum": [
          "high",
          "sweet",
          "lean"
        ],
        "title": "PickTier"
      },
      "PicksResponse": {
        "properties": {
          "picks": {
            "items": {
              "$ref": "#/components/schemas/Pick"
            },
            "type": "array",
            "title": "Picks"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "no_picks_today"
            ],
            "title": "Status"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Checked At"
          },
          "bet_card_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bet Card Date"
          }
        },
        "type": "object",
        "required": [
          "picks",
          "status",
          "last_checked_at"
        ],
        "title": "PicksResponse"
      },
      "SportModel": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "market_groups": {
            "items": {
              "$ref": "#/components/schemas/MarketGroupModel"
            },
            "type": "array",
            "title": "Market Groups"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "market_groups"
        ],
        "title": "SportModel"
      },
      "SportsResponse": {
        "properties": {
          "sports": {
            "items": {
              "$ref": "#/components/schemas/SportModel"
            },
            "type": "array",
            "title": "Sports"
          }
        },
        "type": "object",
        "required": [
          "sports"
        ],
        "title": "SportsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}